Fix: Android: Catch exceptions on multiple SDK initialization causing application crash.
Android SDK throws exception after attempt of initialization of SDK, when it is already initialized. It causes crashes of application, because bridge Javascript->Java doesn't support passing exceptions.
On the other hand, iOS implementation is silent, when SDK is initialized second time.
ReactNative SDK should work the same for both platforms. To unify behaviour of both platform, Android implementation is changed to be silent in this case, similarly to iOS implementation. Only warning log is written.